feat: extracted bb logic to separate module
