+++ functions-python/extract_location/src/main.py
@@ -6,20 +6,23 @@ import uuid
 from datetime import datetime
 
 import functions_framework
-import gtfs_kit
-import numpy
 from cloudevents.http import CloudEvent
-from geoalchemy2 import WKTElement
 from google.cloud import pubsub_v1
+from sqlalchemy import or_
 
 from database_gen.sqlacodegen_models import Gtfsdataset
-from helpers.database import start_db_session
-from helpers.logger import Logger
 from dataset_service.main import (
     DatasetTraceService,
     DatasetTrace,
     Status,
     PipelineStage,
+)
+from helpers.database import start_db_session
+from helpers.logger import Logger
+from .bounding_box_extractor import (
+    get_gtfs_feed_bounds,
+    create_polygon_wkt_element,
+    update_dataset_bounding_box,
 )
 
 logging.basicConfig(level=logging.INFO)
@@ -38,60 +41,6 @@ def parse_resource_data(data: dict) -> tuple:
     bucket_name = data["resource"]["labels"]["bucket_name"]
     url = f"https://storage.googleapis.com/{bucket_name}/{stable_id}/{dataset_id}/{file_name}"
     return stable_id, dataset_id, url
-
-
-def get_gtfs_feed_bounds(url: str, dataset_id: str) -> numpy.ndarray:
-    """
-    Retrieve the bounding box coordinates from the GTFS feed.
-    @:param url (str): URL to the GTFS feed.
-    @:param dataset_id (str): ID of the dataset for logs
-    @:return numpy.ndarray: An array containing the bounds (min_longitude, min_latitude, max_longitude, max_latitude).
-    @:raises Exception: If the GTFS feed is invalid
-    """
-    try:
-        feed = gtfs_kit.read_feed(url, "km")
-        return feed.compute_bounds()
-    except Exception as e:
-        logging.error(f"[{dataset_id}] Error retrieving GTFS feed from {url}: {e}")
-        raise Exception(e)
-
-
-def create_polygon_wkt_element(bounds: numpy.ndarray) -> WKTElement:
-    """
-    Create a WKTElement polygon from bounding box coordinates.
-    @:param bounds (numpy.ndarray): Bounding box coordinates.
-    @:return WKTElement: The polygon representation of the bounding box.
-    """
-    min_longitude, min_latitude, max_longitude, max_latitude = bounds
-    points = [
-        (min_longitude, min_latitude),
-        (min_longitude, max_latitude),
-        (max_longitude, max_latitude),
-        (max_longitude, min_latitude),
-        (min_longitude, min_latitude),
-    ]
-    wkt_polygon = f"POLYGON(({', '.join(f'{lon} {lat}' for lon, lat in points)}))"
-    return WKTElement(wkt_polygon, srid=4326)
-
-
-def update_dataset_bounding_box(session, dataset_id, geometry_polygon):
-    """
-    Update the bounding box of a dataset in the database.
-    @:param session (Session): The database session.
-    @:param dataset_id (str): The ID of the dataset.
-    @:param geometry_polygon (WKTElement): The polygon representing the bounding box.
-    @:raises Exception: If the dataset is not found in the database.
-    """
-    dataset: Gtfsdataset | None = (
-        session.query(Gtfsdataset)
-        .filter(Gtfsdataset.stable_id == dataset_id)
-        .one_or_none()
-    )
-    if dataset is None:
-        raise Exception(f"Dataset {dataset_id} does not exist in the database.")
-    dataset.bounding_box = geometry_polygon
-    session.add(dataset)
-    session.commit()
 
 
 @functions_framework.cloud_event
@@ -241,6 +190,7 @@ def extract_location_batch(_):
     logging.info("Batch function triggered.")
 
     pubsub_topic_name = os.getenv("PUBSUB_TOPIC_NAME", None)
+    force_datasets_update = os.getenv("FORCE_DATASETS_UPDATE", False)
     if pubsub_topic_name is None:
         logging.error("PUBSUB_TOPIC_NAME environment variable not set.")
         return "PUBSUB_TOPIC_NAME environment variable not set.", 500
@@ -251,9 +201,15 @@ def extract_location_batch(_):
     datasets_data = []
     try:
         session = start_db_session(os.getenv("FEEDS_DATABASE_URL"))
+        # Select all latest datasets with no bounding boxes or all datasets if forced
         datasets = (
             session.query(Gtfsdataset)
-            .filter(Gtfsdataset.bounding_box == None)  # noqa: E711
+            .filter(
+                or_(
+                    force_datasets_update,
+                    Gtfsdataset.bounding_box == None,  # noqa: E711
+                )
+            )
             .filter(Gtfsdataset.latest)
             .all()
         )
